Refactor EholdingsTitlesImpl to use RMAPITemplate
Jira: MODKBEKBJ-109
